+++ examples/main.tf
@@ -2,3 +2,48 @@ resource "azurerm_resource_group" "this" {
   name     = uuid()
   location = "westeurope"
 }
+
+module "subnet" {
+  source               = "./module"
+  virtual_network_name = "vn-example"
+  resource_group_name  = azurerm_resource_group.this.name
+  subnets = [
+    {
+      name : "AzureBastionSubnet"
+      address_prefixes : ["10.255.0.0/26"]
+      enforce_private_link_endpoint_network_policies : true
+      enforce_private_link_service_network_policies : false
+      service_endpoints : [
+        "Microsoft.AzureActiveDirectory"
+      ]
+      deligation : {
+        name : null
+        service_delegation : {
+          actions : null
+          name : null
+        }
+      }
+    },
+    {
+      name : "sn-example-endpoints"
+      address_prefixes : ["10.255.0.64/26"]
+      enforce_private_link_endpoint_network_policies : false
+      enforce_private_link_service_network_policies : false
+      service_endpoints : [
+        "Microsoft.AzureActiveDirectory",
+        "Microsoft.Storage",
+      ]
+      deligation : {
+        name : null
+        service_delegation : {
+          actions : null
+          name : null
+        }
+      }
+    }
+  ]
+  tags = {
+    Created : "Terraform",
+    Project : "Simple Example Virtual Network"
+  }
+}