+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-07-25
 
+### Features
+
+- Initial Commit
+- Create ```Subnet (only) for VNET```
+
+
 [Unreleased]: https://github.com/patrickhayo/azr-tf-module-subnet/compare/1.0.0...HEAD
 
 [1.0.0]: https://github.com/patrickhayo/azr-tf-module-subnet/compare/6ca82ab8ccb56c6cf37fa003100592da19bbc285...1.0.0